requirements.txt
moviepy
Pillow
customtkinter
